requirements.txt
# Core Dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-dotenv>=0.19.0
supabase>=0.0.2
openai>=0.27.0
transformers>=4.11.0
torch>=1.9.0
numpy>=1.21.0
pandas>=1.3.0

# Data Processing
python-multipart>=0.0.5
aiofiles>=0.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
pytest-cov>=2.12.1

# Development
black>=21.7b0
isort>=5.9.3
flake8>=3.9.2
mypy>=0.910 